src/ai_agent/requirements_generated.txt
action-msgs==2.0.2
action-tutorials-interfaces==0.33.5
action-tutorials-py==0.33.5
actionlib-msgs==5.3.6
actuator-msgs==0.0.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
ament-cmake-test==2.5.4
ament-copyright==0.17.2
ament-cppcheck==0.17.2
ament-cpplint==0.17.2
ament-flake8==0.17.2
ament-index-python==1.8.1
ament-lint==0.17.2
ament-lint-cmake==0.17.2
ament-package==0.16.4
ament-pep257==0.17.2
ament-uncrustify==0.17.2
ament-xmllint==0.17.2
angles==1.16.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
blockbuster==1.5.24
builtin-interfaces==2.0.2
certifi==2025.7.9
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cloudpickle==3.1.1
composition-interfaces==2.0.2
cryptography==44.0.3
cv-bridge==4.1.0
dataclasses-json==0.6.7
demo-nodes-py==0.33.5
diagnostic-msgs==5.3.6
domain-coordinator==0.12.0
example-interfaces==0.12.0
examples-rclpy-executors==0.19.5
examples-rclpy-minimal-action-client==0.19.5
examples-rclpy-minimal-action-server==0.19.5
examples-rclpy-minimal-client==0.19.5
examples-rclpy-minimal-publisher==0.19.5
examples-rclpy-minimal-service==0.19.5
examples-rclpy-minimal-subscriber==0.19.5
forbiddenfruit==0.1.4
frozenlist==1.7.0
geometry-msgs==5.3.6
gps-msgs==2.0.4
greenlet==3.2.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
idna==3.10
image-geometry==4.1.0
interactive-markers==2.5.4
joint-state-publisher==2.4.0
joint-state-publisher-gui==2.4.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema_rs==0.29.1
langchain==0.3.26
langchain-community==0.3.27
langchain-core==0.3.68
langchain-ollama==0.3.4
langchain-text-splitters==0.3.8
langgraph==0.5.2
langgraph-api==0.2.85
langgraph-checkpoint==2.1.0
langgraph-cli==0.3.4
langgraph-prebuilt==0.5.2
langgraph-runtime-inmem==0.3.4
langgraph-sdk==0.1.72
langsmith==0.4.4
laser-geometry==2.7.0
launch==3.4.4
launch-ros==0.26.7
launch-testing==3.4.4
launch-testing-ros==0.26.7
launch-xml==3.4.4
launch-yaml==3.4.4
lifecycle-msgs==2.0.2
logging-demo==0.33.5
map-msgs==2.4.1
marshmallow==3.26.1
message-filters==4.11.6
multidict==6.6.3
mypy_extensions==1.1.0
nav-msgs==5.3.6
numpy==2.3.1
ollama==0.5.1
orjson==3.10.18
ormsgpack==1.10.0
osrf-pycommon==2.1.6
packaging==24.2
pcl-msgs==1.0.0
pendulum-msgs==0.33.5
propcache==0.3.2
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
PyJWT==2.10.1
python-dotenv==1.1.1
python-qt-binding==2.2.1
PyYAML==6.0.2
qt-dotgraph==2.7.5
qt-gui==2.7.5
qt-gui-cpp==2.7.5
qt-gui-py-common==2.7.5
quality-of-service-demo-py==0.33.5
rcl-interfaces==2.0.2
rclpy==7.1.4
rcutils==6.7.2
requests==2.32.4
requests-toolbelt==1.0.0
resource-retriever==3.4.4
rmw-dds-common==3.1.0
ros-gz-bridge==1.0.11
ros-gz-interfaces==1.0.11
ros-gz-sim==1.0.11
ros2action==0.32.4
ros2bag==0.26.7
ros2bag-mcap-cli==0.26.7
ros2bag-sqlite3-cli==0.26.7
ros2cli==0.32.4
ros2component==0.32.4
ros2doctor==0.32.4
ros2interface==0.32.4
ros2launch==0.26.7
ros2lifecycle==0.32.4
ros2multicast==0.32.4
ros2node==0.32.4
ros2param==0.32.4
ros2pkg==0.32.4
ros2run==0.32.4
ros2service==0.32.4
ros2topic==0.32.4
rosbag2-interfaces==0.26.7
rosbag2-py==0.26.7
rosgraph-msgs==2.0.2
rosidl-adapter==4.6.5
rosidl-cli==4.6.5
rosidl-cmake==4.6.5
rosidl-generator-c==4.6.5
rosidl-generator-cpp==4.6.5
rosidl-generator-py==0.22.1
rosidl-generator-type-description==4.6.5
rosidl-parser==4.6.5
rosidl-pycommon==4.6.5
rosidl-runtime-py==0.13.1
rosidl-typesupport-c==3.2.2
rosidl-typesupport-cpp==3.2.2
rosidl-typesupport-fastrtps-c==3.6.1
rosidl-typesupport-fastrtps-cpp==3.6.1
rosidl-typesupport-introspection-c==4.6.5
rosidl-typesupport-introspection-cpp==4.6.5
rpyutils==0.4.1
rqt-action==2.2.0
rqt-bag==1.5.4
rqt-bag-plugins==1.5.4
rqt-console==2.2.1
rqt-graph==1.5.4
rqt-gui==1.6.0
rqt-gui-py==1.6.0
rqt-msg==1.5.1
rqt-plot==1.4.2
rqt-publisher==1.7.2
rqt-py-common==1.6.0
rqt-py-console==1.2.2
rqt-reconfigure==1.6.2
rqt-service-caller==1.2.1
rqt-shell==1.2.2
rqt-srv==1.2.2
rqt-topic==1.7.3
sensor-msgs==5.3.6
sensor-msgs-py==5.3.6
service-msgs==2.0.2
shape-msgs==5.3.6
sniffio==1.3.1
SQLAlchemy==2.0.41
sros2==0.13.3
sse-starlette==2.1.3
starlette==0.47.1
statistics-msgs==2.0.2
std-msgs==5.3.6
std-srvs==5.3.6
stereo-msgs==5.3.6
structlog==25.4.0
teleop-twist-keyboard==2.4.0
tenacity==9.1.2
tf2-geometry-msgs==0.36.10
tf2-kdl==0.36.10
tf2-msgs==0.36.10
tf2-py==0.36.10
tf2-ros-py==0.36.10
tf2-sensor-msgs==0.36.10
tf2-tools==0.36.10
theora-image-transport==4.0.4
topic-monitor==0.33.5
trajectory-msgs==5.3.6
truststore==0.10.1
turtlesim==1.8.3
type-description-interfaces==2.0.2
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.1
unique-identifier-msgs==2.5.0
urllib3==2.5.0
uvicorn==0.35.0
vision-msgs==4.1.1
visualization-msgs==5.3.6
watchfiles==1.1.0
xacro==2.0.13
xxhash==3.5.0
yarl==1.20.1
zstandard==0.23.0
